renames to ScrollerBase, uses it in Scroller
