+++ src/components/Scroller/Scroller.mdx
@@ -6,9 +6,9 @@ import * as ScrollerStories from './Scroller.stories.svelte';
 
 # Scroller
 
-The `Scroller` component creates a basic scrollytelling graphic with layout options.
+The `Scroller` component creates a basic scrollytelling graphic with layout options 
 
-> This component is designed to handle most common layouts for scrollytelling. To make something more complex, customise [ScrollerBase](https://github.com/reuters-graphics/graphics-components/blob/main/src/components/Scroller/ScrollerBase/index.svelte), which is a Svelte 5 version of the [svelte-scroller](https://github.com/sveltejs/svelte-scroller).
+This component is designed to handle most common layouts for scrollytelling. To make something more complex, customise [ScrollerBase](?path=/story/components-graphics-scrollerbase--docs), which is a Svelte 5 version of the [svelte-scroller](https://github.com/sveltejs/svelte-scroller).
 
 [Demo](?path=/story/components-graphics-scroller--demo)
