--- src/components/SvelteScroller/demo/components/ai2svelte/ai-foreground.svelte
@@ -1,112 +1,0 @@
-<script>
-  // Hard-coding for demo purposes only...
-  import stepXs from './images/Body-issues-key-xs.png';
-  let width = $state();
-</script>
-
-<div id="g-Body-issues-key-box" bind:clientWidth={width}>
-  <!-- Artboard: xs -->
-  {#if width && width >= 0}
-    <div id="g-Body-issues-key-xs" class="g-artboard" style="">
-      <div style="padding: 0 0 48.4848% 0;"></div>
-      <div
-        id="g-Body-issues-key-xs-img"
-        class="g-aiImg"
-        style="background-image: url({stepXs});"
-      ></div>
-      <div
-        id="g-ai0-1"
-        class="g-Layer_1 g-aiAbs g-aiPointText"
-        style="top:19.4775%;margin-top:-10.2px;left:8.1818%;width:276px;"
-      >
-        <p class="g-pstyle0">Likelihood of something happening</p>
-      </div>
-
-      <div
-        id="g-ai0-3"
-        class="g-Layer_1 g-aiAbs g-aiPointText"
-        style="top:55.1025%;margin-top:-10.2px;left:27.2727%;width:68px;"
-      >
-        <p class="g-pstyle1">0-25%</p>
-      </div>
-      <div
-        id="g-ai0-4"
-        class="g-Layer_1 g-aiAbs g-aiPointText"
-        style="top:55.1025%;margin-top:-10.2px;left:74.2424%;width:75px;"
-      >
-        <p class="g-pstyle1">50-75%</p>
-      </div>
-      <div
-        id="g-ai0-5"
-        class="g-Layer_1 g-aiAbs g-aiPointText"
-        style="top:79.4775%;margin-top:-10.2px;left:74.2424%;width:82px;"
-      >
-        <p class="g-pstyle1">75-100%</p>
-      </div>
-      <div
-        id="g-ai0-6"
-        class="g-Layer_1 g-aiAbs g-aiPointText"
-        style="top:83.2275%;margin-top:-10.2px;left:27.2727%;width:77px;"
-      >
-        <p class="g-pstyle1">25-50%</p>
-      </div>
-    </div>
-  {/if}
-</div>
-
-<!-- End ai2html - 2025-03-17 09:52 -->
-
-<!-- Generated by ai2html v0.100.0 - 2025-03-17 09:52 -->
-<!-- ai file: Body-issues-key.ai -->
-<style lang="scss">
-  #g-Body-issues-key-box,
-  #g-Body-issues-key-box .g-artboard {
-    margin: 0 auto;
-  }
-  #g-Body-issues-key-box p {
-    margin: 0;
-  }
-  #g-Body-issues-key-box .g-aiAbs {
-    position: absolute;
-  }
-  #g-Body-issues-key-box .g-aiImg {
-    position: absolute;
-    top: 0;
-    display: block;
-    width: 100% !important;
-    height: 100%;
-    background-size: contain;
-    background-repeat: no-repeat;
-  }
-  #g-Body-issues-key-box .g-aiPointText p {
-    white-space: nowrap;
-  }
-  #g-Body-issues-key-xs {
-    position: relative;
-    overflow: hidden;
-  }
-  #g-Body-issues-key-xs p {
-    font-family: 'Knowledge', 'Source Sans Pro', Arial, sans-serif;
-    font-weight: 500;
-    line-height: 19px;
-    height: auto;
-    opacity: 1;
-    letter-spacing: 0em;
-    font-size: 16px;
-    text-align: left;
-    color: rgb(64, 64, 64);
-    text-transform: none;
-    padding-bottom: 0;
-    padding-top: 0;
-    mix-blend-mode: normal;
-    font-style: normal;
-    position: static;
-  }
-  #g-Body-issues-key-xs .g-pstyle0 {
-    height: 19px;
-  }
-  #g-Body-issues-key-xs .g-pstyle1 {
-    font-weight: 400;
-    height: 19px;
-  }
-</style>
