--- src/components/SvelteScroller/demo/components/ai2svelte/ai-scroller-1.svelte
@@ -1,121 +1,0 @@
-<script lang="ts">
-  // Hard-coding for demo purposes only...
-  import stepXl from './images/step-1-xl.png';
-  import stepLg from './images/step-1-lg.png';
-  import stepMd from './images/step-1-md.png';
-  import stepSm from './images/step-1-sm.png';
-  import stepXs from './images/step-1-xs.png';
-  let width = $state<number>();
-</script>
-
-<!-- Generated by ai2html v0.100.0 - 2021-09-30 14:21 -->
-
-<div id="g-step-1-box" bind:clientWidth={width}>
-  <!-- Artboard: XL -->
-  {#if width && width >= 1200}
-    <div id="g-step-1-xl" class="g-artboard" style="">
-      <div style="padding: 0 0 50% 0;"></div>
-      <div
-        id="g-step-1-xl-img"
-        class="g-aiImg"
-        style="background-image: url({stepXl});"
-      ></div>
-    </div>
-  {/if}
-  <!-- Artboard: LG -->
-  {#if width && width >= 930 && width < 1200}
-    <div id="g-step-1-lg" class="g-artboard" style="">
-      <div style="padding: 0 0 55.3763% 0;"></div>
-      <div
-        id="g-step-1-lg-img"
-        class="g-aiImg"
-        style="background-image: url({stepLg});"
-      ></div>
-    </div>
-  {/if}
-  <!-- Artboard: MD -->
-  {#if width && width >= 660 && width < 930}
-    <div id="g-step-1-md" class="g-artboard" style="">
-      <div style="padding: 0 0 55.7576% 0;"></div>
-      <div
-        id="g-step-1-md-img"
-        class="g-aiImg"
-        style="background-image: url({stepMd});"
-      ></div>
-    </div>
-  {/if}
-  <!-- Artboard: SM -->
-  {#if width && width >= 510 && width < 660}
-    <div id="g-step-1-sm" class="g-artboard" style="">
-      <div style="padding: 0 0 49.6078% 0;"></div>
-      <div
-        id="g-step-1-sm-img"
-        class="g-aiImg"
-        style="background-image: url({stepSm});"
-      ></div>
-    </div>
-  {/if}
-  <!-- Artboard: XS -->
-  {#if width && width >= 0 && width < 510}
-    <div id="g-step-1-xs" class="g-artboard" style="">
-      <div style="padding: 0 0 55.4545% 0;"></div>
-      <div
-        id="g-step-1-xs-img"
-        class="g-aiImg"
-        style="background-image: url({stepXs});"
-      ></div>
-    </div>
-  {/if}
-</div>
-
-<!-- End ai2html - 2021-09-30 14:21 -->
-
-<!-- ai file: step-1.ai -->
-
-<!-- svelte-ignore css_unused_selector -->
-<style lang="scss">
-  #g-step-1-box,
-  #g-step-1-box .g-artboard {
-    margin: 0 auto;
-  }
-  #g-step-1-box p {
-    margin: 0;
-  }
-  #g-step-1-box .g-aiAbs {
-    position: absolute;
-  }
-  #g-step-1-box .g-aiImg {
-    position: absolute;
-    top: 0;
-    display: block;
-    width: 100% !important;
-    height: 100%;
-    background-size: contain;
-    background-repeat: no-repeat;
-  }
-  #g-step-1-box .g-aiPointText p {
-    white-space: nowrap;
-  }
-  #g-step-1-xl {
-    position: relative;
-    overflow: hidden;
-  }
-  #g-step-1-lg {
-    position: relative;
-    overflow: hidden;
-  }
-  #g-step-1-md {
-    position: relative;
-    overflow: hidden;
-  }
-  #g-step-1-sm {
-    position: relative;
-    overflow: hidden;
-  }
-  #g-step-1-xs {
-    position: relative;
-    overflow: hidden;
-  }
-
-  /* Custom CSS */
-</style>
