--- src/components/SvelteScroller/demo/components/ai2svelte/ai-scroller-3.svelte
@@ -1,310 +1,0 @@
-<script lang="ts">
-  // Hard-coding for demo purposes only...
-  // @ts-ignore img
-  import stepXl from './images/step-3-xl.png';
-  // @ts-ignore img
-  import stepLg from './images/step-3-lg.png';
-  // @ts-ignore img
-  import stepMd from './images/step-3-md.png';
-  // @ts-ignore img
-  import stepSm from './images/step-3-sm.png';
-  // @ts-ignore img
-  import stepXs from './images/step-3-xs.png';
-  let width = $state<null | number>(null);
-</script>
-
-<!-- Generated by ai2html v0.100.0 - 2021-09-30 14:28 -->
-
-<div id="g-step-3-box" bind:clientWidth={width}>
-  <!-- Artboard: XL -->
-  {#if width && width >= 1200}
-    <div id="g-step-3-xl" class="g-artboard" style="">
-      <div style="padding: 0 0 50% 0;"></div>
-      <div
-        id="g-step-3-xl-img"
-        class="g-aiImg"
-        style="background-image: url({stepXl});"
-      ></div>
-      <div
-        id="g-ai0-1"
-        class="g-step-3 g-aiAbs g-aiPointText"
-        style="top:19.683%;margin-top:-21.1px;left:54.3441%;width:130px;"
-      >
-        <p class="g-pstyle0">Something</p>
-        <p class="g-pstyle0">happened here</p>
-      </div>
-    </div>
-  {/if}
-  <!-- Artboard: LG -->
-  {#if width && width >= 930 && width < 1200}
-    <div id="g-step-3-lg" class="g-artboard" style="">
-      <div style="padding: 0 0 55.3763% 0;"></div>
-      <div
-        id="g-step-3-lg-img"
-        class="g-aiImg"
-        style="background-image: url({stepLg});"
-      ></div>
-      <div
-        id="g-ai1-1"
-        class="g-step-3 g-aiAbs g-aiPointText"
-        style="top:27.5682%;margin-top:-21px;left:54.9127%;width:118px;"
-      >
-        <p class="g-pstyle0">Something</p>
-        <p class="g-pstyle0">happened here</p>
-      </div>
-    </div>
-  {/if}
-  <!-- Artboard: MD -->
-  {#if width && width >= 660 && width < 930}
-    <div id="g-step-3-md" class="g-artboard" style="">
-      <div style="padding: 0 0 55.7576% 0;"></div>
-      <div
-        id="g-step-3-md-img"
-        class="g-aiImg"
-        style="background-image: url({stepMd});"
-      ></div>
-      <div
-        id="g-ai2-1"
-        class="g-step-3 g-aiAbs g-aiPointText"
-        style="top:23.3631%;margin-top:-21px;left:50.4963%;width:118px;"
-      >
-        <p class="g-pstyle0">Something</p>
-        <p class="g-pstyle0">happened here</p>
-      </div>
-    </div>
-  {/if}
-  <!-- Artboard: SM -->
-  {#if width && width >= 510 && width < 660}
-    <div id="g-step-3-sm" class="g-artboard" style="">
-      <div style="padding: 0 0 49.6078% 0;"></div>
-      <div
-        id="g-step-3-sm-img"
-        class="g-aiImg"
-        style="background-image: url({stepSm});"
-      ></div>
-      <div
-        id="g-ai3-1"
-        class="g-step-3 g-aiAbs g-aiPointText"
-        style="top:20.1486%;margin-top:-21px;left:55.7925%;width:118px;"
-      >
-        <p class="g-pstyle0">Something</p>
-        <p class="g-pstyle0">happened here</p>
-      </div>
-    </div>
-  {/if}
-  <!-- Artboard: XS -->
-  {#if width && width >= 0 && width < 510}
-    <div id="g-step-3-xs" class="g-artboard" style="">
-      <div style="padding: 0 0 55.4545% 0;"></div>
-      <div
-        id="g-step-3-xs-img"
-        class="g-aiImg"
-        style="background-image: url({stepXs});"
-      ></div>
-      <div
-        id="g-ai4-1"
-        class="g-step-3 g-aiAbs g-aiPointText"
-        style="top:42.0634%;margin-top:-21px;left:27.3523%;width:118px;"
-      >
-        <p class="g-pstyle0">Something</p>
-        <p class="g-pstyle0">happened here</p>
-      </div>
-    </div>
-  {/if}
-</div>
-
-<!-- End ai2html - 2021-09-30 14:28 -->
-
-<!-- ai file: step-3.ai -->
-
-<style lang="scss">
-  #g-step-3-box,
-  #g-step-3-box .g-artboard {
-    margin: 0 auto;
-  }
-  #g-step-3-box p {
-    margin: 0;
-  }
-  #g-step-3-box .g-aiAbs {
-    position: absolute;
-  }
-  #g-step-3-box .g-aiImg {
-    position: absolute;
-    top: 0;
-    display: block;
-    width: 100% !important;
-    height: 100%;
-    background-size: contain;
-    background-repeat: no-repeat;
-  }
-  #g-step-3-box .g-aiPointText p {
-    white-space: nowrap;
-  }
-  #g-step-3-xl {
-    position: relative;
-    overflow: hidden;
-  }
-  #g-step-3-xl p {
-    font-family:
-      'Source Sans Pro',
-      -apple-system,
-      BlinkMacSystemFont,
-      'Segoe UI',
-      Roboto,
-      'Helvetica Neue',
-      Helvetica,
-      Arial,
-      sans-serif;
-    font-weight: 400;
-    line-height: 20px;
-    height: auto;
-    opacity: 1;
-    letter-spacing: 0em;
-    font-size: 18px;
-    text-align: left;
-    color: rgb(0, 0, 0);
-    text-transform: none;
-    padding-bottom: 0;
-    padding-top: 0;
-    mix-blend-mode: normal;
-    font-style: normal;
-    position: static;
-  }
-  #g-step-3-xl .g-pstyle0 {
-    height: 20px;
-  }
-  #g-step-3-lg {
-    position: relative;
-    overflow: hidden;
-  }
-  #g-step-3-lg p {
-    font-family:
-      'Source Sans Pro',
-      -apple-system,
-      BlinkMacSystemFont,
-      'Segoe UI',
-      Roboto,
-      'Helvetica Neue',
-      Helvetica,
-      Arial,
-      sans-serif;
-    font-weight: 400;
-    line-height: 20px;
-    height: auto;
-    opacity: 1;
-    letter-spacing: 0em;
-    font-size: 16px;
-    text-align: left;
-    color: rgb(0, 0, 0);
-    text-transform: none;
-    padding-bottom: 0;
-    padding-top: 0;
-    mix-blend-mode: normal;
-    font-style: normal;
-    position: static;
-  }
-  #g-step-3-lg .g-pstyle0 {
-    height: 20px;
-  }
-  #g-step-3-md {
-    position: relative;
-    overflow: hidden;
-  }
-  #g-step-3-md p {
-    font-family:
-      'Source Sans Pro',
-      -apple-system,
-      BlinkMacSystemFont,
-      'Segoe UI',
-      Roboto,
-      'Helvetica Neue',
-      Helvetica,
-      Arial,
-      sans-serif;
-    font-weight: 400;
-    line-height: 20px;
-    height: auto;
-    opacity: 1;
-    letter-spacing: 0em;
-    font-size: 16px;
-    text-align: left;
-    color: rgb(0, 0, 0);
-    text-transform: none;
-    padding-bottom: 0;
-    padding-top: 0;
-    mix-blend-mode: normal;
-    font-style: normal;
-    position: static;
-  }
-  #g-step-3-md .g-pstyle0 {
-    height: 20px;
-  }
-  #g-step-3-sm {
-    position: relative;
-    overflow: hidden;
-  }
-  #g-step-3-sm p {
-    font-family:
-      'Source Sans Pro',
-      -apple-system,
-      BlinkMacSystemFont,
-      'Segoe UI',
-      Roboto,
-      'Helvetica Neue',
-      Helvetica,
-      Arial,
-      sans-serif;
-    font-weight: 400;
-    line-height: 20px;
-    height: auto;
-    opacity: 1;
-    letter-spacing: 0em;
-    font-size: 16px;
-    text-align: left;
-    color: rgb(0, 0, 0);
-    text-transform: none;
-    padding-bottom: 0;
-    padding-top: 0;
-    mix-blend-mode: normal;
-    font-style: normal;
-    position: static;
-  }
-  #g-step-3-sm .g-pstyle0 {
-    height: 20px;
-  }
-  #g-step-3-xs {
-    position: relative;
-    overflow: hidden;
-  }
-  #g-step-3-xs p {
-    font-family:
-      'Source Sans Pro',
-      -apple-system,
-      BlinkMacSystemFont,
-      'Segoe UI',
-      Roboto,
-      'Helvetica Neue',
-      Helvetica,
-      Arial,
-      sans-serif;
-    font-weight: 400;
-    line-height: 20px;
-    height: auto;
-    opacity: 1;
-    letter-spacing: 0em;
-    font-size: 16px;
-    text-align: left;
-    color: rgb(0, 0, 0);
-    text-transform: none;
-    padding-bottom: 0;
-    padding-top: 0;
-    mix-blend-mode: normal;
-    font-style: normal;
-    position: static;
-  }
-  #g-step-3-xs .g-pstyle0 {
-    height: 20px;
-  }
-
-  /* Custom CSS */
-</style>
